Renamed AlertFactionCombatMetrics
